fix(chat): cleanups
